(maint) Modularize the admin helpers and general querying
This commit moves the `puppet-db` helper code (admin tasks) to its own
namespace `puppetdb::admin` and moves the `execute_query` function to be
a method on a `puppetdb::client::Config` instead.
